move weighted_gmm_pdf
